+++ docker/services/rightsizer_application_service.py
@@ -5,7 +5,7 @@ from modular_sdk.services.customer_service import CustomerService
 from pynamodb.attributes import MapAttribute
 
 from commons.constants import MAESTRO_RIGHTSIZER_APPLICATION_TYPE, \
-    MAESTRO_RIGHTSIZER_LICENSES_APPLICATION_TYPE
+    MAESTRO_RIGHTSIZER_LICENSES_APPLICATION_TYPE, TENANTS_ATTR
 from models.application_attributes import (RightsizerApplicationMeta,
                                            RightsizerLicensesApplicationMeta)
 
@@ -59,3 +59,21 @@ class RightSizerApplicationService(ApplicationService):
         else:
             application_meta_obj = meta_attr_class()
         return application_meta_obj
+
+    def get_by_license_key(self, customer, license_key: str):
+        applications = self.list(
+            customer=customer,
+            _type=MAESTRO_RIGHTSIZER_LICENSES_APPLICATION_TYPE,
+            deleted=False
+        )
+        for application in applications:
+            app_meta = self.get_application_meta(application=application)
+            if app_meta.license_key == license_key:
+                return application
+
+    def list_allowed_license_tenants(self, application: Application):
+        app_meta = self.get_application_meta(application=application)
+        customer_map = app_meta.customers.get(application.customer_id)
+        if not customer_map:
+            return
+        return list(customer_map.get(TENANTS_ATTR, []))