+++ docker/services/service_provider.py
@@ -78,7 +78,6 @@ class ServiceProvider:
         __token_service = None
         __license_manager_service = None
         __key_management_service = None
-        __license_service = None
 
         def __str__(self):
             return id(self)
@@ -335,15 +334,6 @@ class ServiceProvider:
                     key_management_client=self.standalone_key_management()
                 )
             return self.__key_management_service
-
-        def license_service(self):
-            if not self.__license_service:
-                from services.license_service import \
-                    LicenseService
-                self.__license_service = LicenseService(
-                    settings_service=self.settings_service()
-                )
-            return self.__license_service
 
     instance = None