+++ r8s/r8scli/group/license.py
@@ -8,40 +8,14 @@ def license():
     """Manages RIGHTSIZER License Entity"""
 
 
-@license.command(cls=ViewCommand, name='describe')
-@click.option('--license_key', '-lk', type=str, required=False,
-              help='License key to describe.')
-@cli_response()
-def describe(license_key=None):
-    """
-    Describes a RIGHTSIZER License.
-    """
-    from r8scli.service.initializer import init_configuration
-    return init_configuration().license_get(
-        license_key=license_key)
-
-
-@license.command(cls=ViewCommand, name='delete')
-@click.option('--license_key', '-lk', type=str, required=True,
-              help='License key to delete.')
-@cli_response()
-def delete(license_key=None):
-    """
-    Deletes a RIGHTSIZER License.
-    """
-    from r8scli.service.initializer import init_configuration
-    return init_configuration().license_delete(
-        license_key=license_key)
-
-
 @license.command(cls=ViewCommand, name='sync')
-@click.option('--license_key', '-lk', type=str, required=True,
-              help='License key to synchronize.')
+@click.option('--application_id', '-aid', type=str, required=True,
+              help='RIGHTSIZER_LICENSES application to sync license.')
 @cli_response()
-def sync(license_key=None):
+def sync(application_id: str):
     """
     Synchronizes a RIGHTSIZER License.
     """
     from r8scli.service.initializer import init_configuration
     return init_configuration().license_sync_post(
-        license_key=license_key)
+        application_id=application_id)