+++ src/lambdas/r8s_api_handler/handler.py
@@ -22,8 +22,6 @@ from lambdas.r8s_api_handler.processors.license_manager_client_processor import 
     LicenseManagerClientProcessor
 from lambdas.r8s_api_handler.processors.license_manager_config_processor import \
     LicenseManagerConfigProcessor
-from lambdas.r8s_api_handler.processors.license_processor import \
-    LicenseProcessor
 from lambdas.r8s_api_handler.processors.license_sync_processor import \
     LicenseSyncProcessor
 from lambdas.r8s_api_handler.processors.mail_report_processor import \
@@ -67,7 +65,6 @@ from services.environment_service import EnvironmentService
 from services.job_service import JobService
 from services.key_management_service import KeyManagementService
 from services.license_manager_service import LicenseManagerService
-from services.license_service import LicenseService
 from services.rbac.access_control_service import AccessControlService
 from services.rbac.iam_service import IamService
 from services.recommendation_history_service import \
@@ -142,8 +139,7 @@ class R8sApiHandler(AbstractApiHandlerLambda):
                  customer_preferences_service: CustomerPreferencesService,
                  resize_service: ResizeService,
                  key_management_service: KeyManagementService,
-                 license_manager_service: LicenseManagerService,
-                 license_service: LicenseService):
+                 license_manager_service: LicenseManagerService):
         self.user_service = user_service
         self.access_control_service = access_control_service
         self.algorithm_service = algorithm_service
@@ -170,7 +166,6 @@ class R8sApiHandler(AbstractApiHandlerLambda):
         self.resize_service = resize_service
         self.key_management_service = key_management_service
         self.license_manager_service = license_manager_service
-        self.license_service = license_service
 
         self.processor_registry = {
             SIGNIN_ACTION: self._instantiate_signin_processor,
@@ -204,7 +199,6 @@ class R8sApiHandler(AbstractApiHandlerLambda):
                 self._instantiate_resize_insights_processor,
             LM_SETTING_CONFIG_ACTION: self._instantiate_lm_config_processor,
             LM_SETTING_CLIENT_ACTION: self._instantiate_lm_client_processor,
-            LICENSE_ACTION: self._instantiate_license_processor,
             LICENSE_SYNC_ACTION: self._instantiate_license_sync_processor
         }
 
@@ -290,7 +284,6 @@ class R8sApiHandler(AbstractApiHandlerLambda):
             customer_service=self.customer_service,
             application_service=self.application_service,
             parent_service=self.parent_service,
-            license_service=self.license_service,
             license_manager_service=self.license_manager_service
         )
 
@@ -312,7 +305,6 @@ class R8sApiHandler(AbstractApiHandlerLambda):
             shape_service=self.shape_service,
             shape_price_service=self.shape_price_service,
             parent_service=self.parent_service,
-            license_service=self.license_service,
             license_manager_service=self.license_manager_service
         )
 
@@ -358,7 +350,6 @@ class R8sApiHandler(AbstractApiHandlerLambda):
             application_service=self.application_service,
             parent_service=self.parent_service,
             tenant_service=self.tenant_service,
-            license_service=self.license_service,
             license_manager_service=self.license_manager_service
         )
 
@@ -438,17 +429,11 @@ class R8sApiHandler(AbstractApiHandlerLambda):
             key_management_service=self.key_management_service
         )
 
-    def _instantiate_license_processor(self):
-        return LicenseProcessor(
-            license_service=self.license_service,
-            algorithm_service=self.algorithm_service
-        )
-
     def _instantiate_license_sync_processor(self):
         return LicenseSyncProcessor(
-            license_service=self.license_service,
             license_manager_service=self.license_manager_service,
-            algorithm_service=self.algorithm_service
+            algorithm_service=self.algorithm_service,
+            application_service=self.application_service
         )
 
 
@@ -478,8 +463,7 @@ HANDLER = R8sApiHandler(
     customer_preferences_service=SERVICE_PROVIDER.customer_preferences_service(),
     resize_service=SERVICE_PROVIDER.resize_service(),
     key_management_service=SERVICE_PROVIDER.key_management_service(),
-    license_manager_service=SERVICE_PROVIDER.license_manager_service(),
-    license_service=SERVICE_PROVIDER.license_service()
+    license_manager_service=SERVICE_PROVIDER.license_manager_service()
 )