+++ src/models/application_attributes.py
@@ -1,6 +1,10 @@
 from pynamodb.attributes import MapAttribute, UnicodeAttribute, \
     NumberAttribute, ListAttribute
 
+PERMITTED_ATTACHMENT = 'permitted'
+PROHIBITED_ATTACHMENT = 'prohibited'
+ALLOWED_ATTACHMENT_MODELS = (PERMITTED_ATTACHMENT, PERMITTED_ATTACHMENT)
+
 
 class ConnectionAttribute(MapAttribute):
     host = UnicodeAttribute(null=True)
@@ -15,11 +19,21 @@ class RightsizerApplicationMeta(MapAttribute):
     connection = ConnectionAttribute(null=True)
 
 
+class AllowanceAttribute(MapAttribute):
+    time_range = UnicodeAttribute(null=True)
+    job_balance = NumberAttribute(null=True)
+    balance_exhaustion_model = UnicodeAttribute(null=True)
+
+
 class RightsizerLicensesApplicationMeta(MapAttribute):
     cloud = UnicodeAttribute(null=True)
     algorithm_map = MapAttribute(null=True)
     license_key = UnicodeAttribute(null=True)
-    tenants = ListAttribute(of=UnicodeAttribute, default=[])
+    tenant_license_key = UnicodeAttribute(null=True)
+    expiration = UnicodeAttribute(null=True)
+    latest_sync = UnicodeAttribute(null=True)
+    allowance = AllowanceAttribute(null=True)
+    customers = MapAttribute(null=True)
 
 
 class RightSizerDojoApplicationMeta(MapAttribute):