+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     * `remove_tags` - Removes specified resource groups from group configuration
     * `add_resource_group` - Adds specified resource groups to configuration
     * `remove_resource_group` - Removes specified resource groups from configuration
+* Delete `Licenses` collection and related crud commands, move all license-related data to RIGHTSIZER_LICENSES application
 
 ## [3.12.0] - 2025-03-14
 * Add ability to run Syndicate RightSizer instance with SaaS ModularSDK / S3